Refactored computer_core functions
